refactor: Добавлен вспомогательный модуль для работы с ДОМом
