+++ js/render-advertisements.js
@@ -1,23 +1,6 @@
 import { generateData } from './mock-data.js';
-import { toggleClassHidden } from './util.js';
+import { toggleClassHidden, QuerySelector } from './dom-util.js';
 
-const QuerySelector = {
-  ADS_CONTAINER: '.map__canvas',
-  ADS_TEMPLATE: '#card',
-  AD: '.popup',
-  AVATAR: '.popup__avatar',
-  TITLE: '.popup__title',
-  ADDRESS: '.popup__text--address',
-  PRICE: '.popup__text--price',
-  TYPE: '.popup__type',
-  CAPASITY: '.popup__text--capacity',
-  TIME: '.popup__text--time',
-  FEATURE: '.popup__feature',
-  FEATURE_CONTAINER: '.popup__features',
-  DESCRIPTION: '.popup__description',
-  PHOTO_CONTAINER: '.popup__photos',
-  PHOTO: '.popup__photo',
-};
 const typeToRus = {
   flat: 'Квартира',
   bungalow: 'Бунгало',
